Strides-main/requirement.txt
streamlit
google-genai
google-generativeai
requests
python-dotenv
